refactor [math] - torch: create a package for torch
